+++ config.c
@@ -953,6 +953,9 @@ parse_section_main(struct context *ctx)
     else if (strcmp(key, "minimal-lines") == 0)
         return value_to_bool(ctx, &conf->minimal_lines);
 
+    else if (strcmp(key, "hide-prompt") == 0)
+        return value_to_bool(ctx, &conf->hide_prompt);
+
     else if (strcmp(key, "width") == 0)
         return value_to_uint32(ctx, 10, &conf->chars);